Changes:
Global Cache: A global dictionary reverse_memory_Cache has been added to store the reverse task-to-UUID mapping in memory.

Initialisation: The cache is now populated once at the start of the program by the reverse_memory_cache() function.

In-Memory Lookups [ not yet added] : Functions like remove_task() and update_task() will now be updated to use the fast, in-memory cache for all lookups, instead of reading from the JSON file on every call.